Fix(quality): address review comments
