+++ flujo/domain/consensus.py
@@ -55,7 +55,7 @@ def majority_vote(results: Sequence[StepResult]) -> StepResult:
         if key not in first_index:
             first_index[key] = idx
 
-    # Tie-break on earliest appearance to keep results stable.
+    # Tie-break on earliest appearance (negative index makes max prefer smaller idx).
     best_key = max(counts.keys(), key=lambda k: (len(counts[k]), -first_index[k]))
     winners = counts[best_key]
     return _clone_with_meta(
@@ -91,7 +91,7 @@ def code_consensus(results: Sequence[StepResult]) -> StepResult:
         if key not in first_index:
             first_index[key] = idx
 
-    # Tie-break on earliest appearance to keep results stable.
+    # Tie-break on earliest appearance (negative index makes max prefer smaller idx).
     best_key = max(counts.keys(), key=lambda k: (len(counts[k]), -first_index[k]))
     winners = counts[best_key]
     if len(winners) <= 1:
